backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
Werkzeug==3.1.3
zipp==3.21.0
